+++ moabet/src/main/java/com/api/moabet/dto/event/EventRequestDTO.java
@@ -1,15 +1,17 @@
 package com.api.moabet.dto.event;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record EventRequestDTO(
 
-        @NotBlank(message = "Name is required") String name,
+                @NotBlank(message = "Name is required") String name,
 
-        String description,
+                String description,
 
-        @NotNull(message = "Odds is required") @DecimalMin(value = "1.01", message = "Odds must be greater than 1.0") Double odds) {
+                @NotNull(message = "Odds is required") @DecimalMin(value = "1.01", message = "Odds must be greater than 1.0") BigDecimal odds) {
 
 }
